feat: refactor fakers to simplify structure and improve usability; update launch settings schema URL
