+++ social-campus-server/Infrastructure/Data/DatabaseSeeder.cs
@@ -6,12 +6,7 @@ using Domain.Models.FollowModel;
 using Domain.Models.PublicationLikeModel;
 using Domain.Models.PublicationModel;
 using Domain.Models.UserModel;
-using Infrastructure.Fakers.CommentFaker;
-using Infrastructure.Fakers.CommentLikeFaker;
-using Infrastructure.Fakers.FollowFaker;
-using Infrastructure.Fakers.PublicationFaker;
-using Infrastructure.Fakers.PublicationLikeFaker;
-using Infrastructure.Fakers.UserFaker;
+using Infrastructure.Fakers;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.DependencyInjection;
 
@@ -27,27 +22,27 @@ public static class DatabaseSeeder
         Randomizer.Seed = new Random(42);
 
         var passwordHasher = serviceCollection.BuildServiceProvider().GetRequiredService<IPasswordHasher>();
-        var users = await UserFakerFactory.GenerateAsync(20, passwordHasher);
+        var users = await UserFaker.GenerateAsync(20, passwordHasher);
         await context.Set<User>().AddRangeAsync(users, ct);
         await context.SaveChangesAsync(ct);
 
-        var publications = PublicationFakerFactory.Generate(60, users);
+        var publications = PublicationFaker.Generate(60, users);
         await context.Set<Publication>().AddRangeAsync(publications, ct);
         await context.SaveChangesAsync(ct);
 
-        var comments = CommentFakerFactory.Generate(150, users, publications);
+        var comments = CommentFaker.Generate(150, users, publications);
         await context.Set<Comment>().AddRangeAsync(comments, ct);
         await context.SaveChangesAsync(ct);
 
-        var commentLikes = CommentLikeFakerFactory.Generate(users, comments);
+        var commentLikes = CommentLikeFaker.Generate(users, comments);
         await context.Set<CommentLike>().AddRangeAsync(commentLikes, ct);
         await context.SaveChangesAsync(ct);
 
-        var publicationLikes = PublicationLikeFakerFactory.Generate(users, publications);
+        var publicationLikes = PublicationLikeFaker.Generate(users, publications);
         await context.Set<PublicationLike>().AddRangeAsync(publicationLikes, ct);
         await context.SaveChangesAsync(ct);
 
-        var follows = FollowFakerFactory.Generate(users);
+        var follows = FollowFaker.Generate(users);
         await context.Set<Follow>().AddRangeAsync(follows, ct);
         await context.SaveChangesAsync(ct);
     }
