+++ social-campus-server/Presentation/Properties/launchSettings.json
@@ -1,5 +1,5 @@
 {
-  "$schema": "http://json.schemastore.org/launchsettings.json",
+  "$schema": "https://json.schemastore.org/launchsettings.json",
   "profiles": {
     "Development": {
       "commandName": "Project",
